McpClient/requirements.txt
autogen_agentchat==0.5.1
autogen_core==0.5.1
autogen_ext==0.5.1
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.44.1
